+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - `e t` to edit entry task
 - `e p` to edit entry project
 - Simple calculations in time entry
+- `reaper-get-running-timer-note` for getting note of running timer.
 
 ### Changed
 - `reaper-edit-entry` moved from `e` to `e e`